added connection to firebase cloud and updating the delete feature

+++ server/controller/handleEvent.js
@@ -1,10 +1,149 @@
 const {PrismaClient} = require('@prisma/client');
 const { all } = require('axios');
-
-
-
-
+const fs = require('fs')
+const path = require('path')
+const {getStorage, ref, deleteObject, getDownloadURL, UuploadBytesResumable, uploadBytesResumable}=require("firebase/storage")
+const {initializeApp} = require("firebase/app")
+const config = require('../config/firebase')
+const Imageupload = require("../Middlware/coverImage");
+const giveCurrentDateTime = require('../utils/date')
 const prisma = new PrismaClient()
+
+
+// initialize firebase application
+initializeApp(config.firebaseConfig);
+
+// initialize the storage with the firebase service 
+const storage = getStorage();
+
+exports.createEvent = async(req,res) => {
+
+    
+    
+        try{
+            if(!req.file){
+                res.status(400).json({message:"File could not be found"})
+            }
+            // storage path reference 
+            const dateTime = giveCurrentDateTime();
+            const storageRef = ref(storage, `files/${dateTime}_${req.file.originalname}`)
+            
+
+            const metadata = {
+                contentType: req.file.mimetype
+            }
+
+            // upload the file to the firebase storage
+            const uploadaction = uploadBytesResumable(storageRef, req.file.buffer, metadata)
+        
+            // wait for the upload to complete 
+            const snapshot = await uploadaction;
+
+            // gets the url of the post 
+            const downloadImageUrl = await getDownloadURL(snapshot.ref)
+
+            res.status(200).json({ message:"file successful uploaded" });
+
+
+        // Converts the eventtype to an int
+        const stringEventType = req.body.eventType
+        IntEventType = parseInt(stringEventType, 10)
+        const userId = req.user.userId
+        console.log(downloadImageUrl);
+
+        // Stores the data in the Database 
+        const newEventBody = {
+            ...req.body, 
+            ImageCoverUpload: downloadImageUrl,
+            eventType: IntEventType,
+            eventHost: userId,
+            
+        }
+
+        
+    
+        try {
+            const newCreateEvent = await prisma.eventPrompt.create({ data: newEventBody});
+            console.log(newCreateEvent, "successful uploaded")
+            // res.status(200).json({messaage: "successful uploaded on the database"})
+            
+            
+        } catch (error) {
+            console.error(error)
+            res.status(500).send('issue server side')
+        }
+
+
+            // res.send({
+            //     messaage: 'uploaded file to firebase storage',
+            //     name: req.file.originalname,
+            //     type: req.file.mimetype,
+            //     downloadImageUrl: downloadImageUrl
+            // })
+            // res.status(200).json({downloadImageUrl})
+
+            //  gets the dowload url after the upload is completed
+            // uploadaction.on('state_changed', 
+            // (snapshot) => {
+            //     // handle the upload progress, like loading bar 
+            //     const progress = (snapshot.bytesTransferred / snapshot.totalBytes) * 100
+            //     console.log(progress  );
+            // }, (error) => {
+            //     console.error(error)
+            //     res.status(500).json({error: "Error uploading File"})
+            // }, 
+            // async() => {
+            //     const downloadImageUrl = await getDownloadURL(uploadaction.ref)
+            //     console.log(downloadImageUrl);
+            //     res.status(200).json({downloadImageUrl})
+            // }
+            // )
+    
+        }catch(error){
+
+            console.log(error)
+            res.status(500).json({message:"unexpected Error, trying to handle the file data"})
+        }
+
+        
+        // res.json(newEventBody)
+    }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+// find the Event Object throw the user id and respond with the object 
 
 
 exports.findEvents = async(req,res) => {
@@ -37,49 +176,48 @@ exports.findEvents = async(req,res) => {
 }
 
 
+
+
+
+
+
+
+// Delete an Event Object 
+
 exports.deleteEvent = async(req,res) => {
     const id = req.params.id;
     console.log(id)
-
-
-    const deletedEvent = await prisma.eventPrompt.delete({
-        where: {id: id},
-    })
-
-    res.json(deletedEvent)
-}
-
-
-exports.createEvent = async(req,res) => {
-  
-//   console.log(req.body)
-//    console.log(req.user)
-
-    const stringEventType = req.body.eventType
-    IntEventType = parseInt(stringEventType, 10)
-    
-    const userId = req.user.userId
-  
-    
-    const newEventBody = {
-        ...req.body, 
-        ImageCoverUpload: req.file.filename,
-        eventType: IntEventType,
-        eventHost: userId,
+    console.log(req.body.eventpath);
+    const imagePath = req.body.eventpath
+    // console.log(req.params);
+    try{
         
-    }
+        if(!id){
+            res.status(400).json({message: 'Image could not be found, provoke bad request'})
+        }else{
+            const deletedEvent = await prisma.eventPrompt.delete({
+                where: {id: id},
+            })
+            console.log('Event is successfull deleted from the db ');
+            res.status(200).json({message: "Event is successfull deleted from the db "})
+        }
 
-
+        const storageRef = ref(storage, imagePath)
+        await deleteObject(storageRef)
+        
+        console.log('Image is deleted')
     
-    // console.log('test')
-    try {
-        const newCreateEvent = await prisma.eventPrompt.create({ data: newEventBody});
-        console.log(newCreateEvent)
+       
 
         
-    } catch (error) {
-        console.error(error)
-        res.status(500).send('issue server side')
+    }catch(error){
+        res.status(500).json({message:"Error trying to Delete the object"})
     }
-    res.json(newEventBody)
+
+    // const deletedEvent = await prisma.eventPrompt.delete({
+    //     where: {id: id},
+    // })
+
+    // res.json(deletedEvent)
 }
+
